UBUNTU: SAUCE: (drop after 3.3) ALSA: HDA: Remove quirk for Gigabyte GA-H61M-S2PV
This model quirk causes codec parsing to fail for the bug reporter.
As a result, alsamixer cannot even be opened.

Verification: I have tested the changed code in hda-emu and then
the error disappears (and no new errors are appearing), and hda-emu
is a safe way to test/verify these types of errors.

Upstream: has already removed this model completely as part of a
bigger rewrite.

BugLink: https://bugs.launchpad.net/bugs/948360
Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>