Revert "x86/ioapic: Add register level checks to detect bogus io-apic entries"
BugLink: http://bugs.launchpad.net/bugs/974982

This reverts commit 73d63d038ee9f769f5e5b46792d227fe20e442c5.

This patch results in a NULL pointer dereference which renders a system
unbootable if attempting to boot in pvops mode.  We received this patch
via upstream stable v3.2.14.  However, upstream discussion indicates
this patch was not meant to be included in any v3.2.y upstream stable
release.  Revert it.

http://www.mail-archive.com/stable@vger.kernel.org/msg04782.html

Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>