UBUNTU: SAUCE: ite-cir: postpone ISR registration
BugLink: http://bugs.launchpad.net/bugs/984387

An early registration of an ISR was causing a crash to several users.  The
reason was that IRQs were being triggered before the driver initialisation
was completed.

This patch fixes this by moving the invocation to request_irq() to a later
stage on the driver probe function.

Signed-off-by: Luis Henriques <luis.henriques@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Seth Forshee <seth.forshee@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>